x/dataocean/types/message_submit_pay_sign_test.go
package types

import (
	"testing"

	"dataocean/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSubmitPaySign_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubmitPaySign
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubmitPaySign{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubmitPaySign{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
